requirements.txt
matplotlib==3.2.2
numpy==1.18.5
pandas==1.0.5
requests==2.24.0
scikit-learn==0.23.1
pickleshare==0.7.5
seaborn==0.10.1
google_images_download==2.8.0
pillow==8.0.1